Various changes
Added try except blocks to the tokenizer to make a couple of the common parsing errors easier to understand.

Had to change some things in repair_metabolites_matching because the pyarrow dtype pandas regex matching methods don't behave the same as the python regex methods.

Lots of work in repair on getting the metabolite name matching to be more robust, and removal of duplicates. Added methods to the mwtab class to set some internal data members from a dataframe, so that repair isn't doing it directly.