Added to validator
Added some of the todos to the validator.
Needed to investigate one of the todos before committing to it, and ended up finding an issue with some JSON outputs not being valid JSON. Tracked that down and fixed it in mwtab.py and tested it. All mwtab files parsed and saved out to JSON were valid JSON.
Also went through the standard column names and made them all lower case and using underscores. We had decided to do that, but I hadn't gotten around to it.