Many updates
Forgot to commit as I made major changes. The biggest changes were changing DuplicatesDict and updating the MWTabFile class so that the text and JSON representations are the same when read in. There were many changes even into the tokenizer to support the internal representation changes. The DuplicatesDict changes were a part of this as well, just making it much faster than it was previously.

Changes in validator to remove some code that I changed over to test table wise instead of looping over the list of dicts. Also some of the updates to the mwschema and propagating them through the package are in this commit.